CMPE 321 Database/Project-3/src/321project3/requirements.txt
asgiref==3.5.0
Django==4.0.4
django-crispy-forms==1.14.0
mysql-connector-python==8.0.29
mysqlclient==2.1.0
protobuf==3.20.1
sqlparse==0.4.2
